requirements.txt
pathway>=0.13.0
google-generativeai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
